Sync theme-color metadata across all routes
